requirements.txt
langchain>=0.2
openai>=1.30
scikit-learn
jinja2
pyyaml
python-dotenv
